fix: store accel data as m/s² (float) in CSV instead of raw ADC counts
- rec_row_t.accel_x/y/z: int16_t → float, units m/s²
- ecg_sampler_task: row.accel = raw / 16384.0f * 9.81f at point of capture
- svc_record.c: format spec %d → %.4f for three accel columns; CSV header
  renamed accel_x/y/z → accel_x_ms2/accel_y_ms2/accel_z_ms2

Ring buffers remain int32_t raw chip values — chart autoscale is
unit-agnostic, no display change.

+++ main/app_state.h
@@ -13,7 +13,7 @@ typedef struct {
     int32_t  ecg3;             /* Lead III (ADS1293 CH3) */
     int32_t  ppg;              /* int32_t: holds 19-bit MAX86140 samples without truncation */
     int16_t  resp, nas, fcg1;
-    int16_t  accel_x, accel_y, accel_z;  /* MPU-6050 raw; divide by 16384 × 9.81 for m/s² */
+    float    accel_x, accel_y, accel_z;  /* MPU-6050, m/s² (±2g = ±19.62 m/s²) */
     int32_t  drift_ms;
     uint8_t  batt_pct;  /* E-Gauge percent, written every ~30 s, 0 between updates */
     uint8_t  spo2;
